self cr fixes

+++ Assets/Core/Game/Domains/GamePlay/Simulation/Match/Scripts/Stage/IStageDataService.cs
@@ -11,7 +11,7 @@ namespace Core.Game.Domains.GamePlay.Simulation.Match.Scripts.Stage
         float StageRestartTimer { get; set; }
         void ClearData();
         Dictionary<ushort, int> GemsCollectedPerTeam { get; }
-        void AddGemsForTeam(ushort teamAlive, ushort gemsDelta);
+        void AddGemsForTeam(ushort teamAlive, int gemsDelta);
         void InitEntryPoint();
     }
 }
